Enhance unit arithmetic and tests + Bug fixes

+++ qtty-ffi/src/lib.rs
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+// Copyright (C) 2026 Vallés Puig, Ramon
+
 //! C-compatible FFI bindings for `qtty` physical quantities and unit conversions.
 //!
 //! `qtty-ffi` provides a stable C ABI for `qtty`, enabling interoperability with C/C++ code
@@ -187,7 +190,7 @@ mod tests {
         let status = unsafe { qtty_quantity_convert(src, UnitId::Kilometer as u32, &mut out) };
         assert_eq!(status, QttyStatus::Ok);
         assert!((out.value - 1.0).abs() < 1e-12);
-        assert_eq!(out.unit, UnitId::Kilometer);
+        assert_eq!(out.unit, UnitId::Kilometer as u32);
     }
 
     #[test]
@@ -197,7 +200,7 @@ mod tests {
         let status = unsafe { qtty_quantity_convert(src, UnitId::Hour as u32, &mut out) };
         assert_eq!(status, QttyStatus::Ok);
         assert!((out.value - 1.0).abs() < 1e-12);
-        assert_eq!(out.unit, UnitId::Hour);
+        assert_eq!(out.unit, UnitId::Hour as u32);
     }
 
     #[test]
@@ -208,7 +211,7 @@ mod tests {
         let status = unsafe { qtty_quantity_convert(src, UnitId::Radian as u32, &mut out) };
         assert_eq!(status, QttyStatus::Ok);
         assert!((out.value - PI).abs() < 1e-12);
-        assert_eq!(out.unit, UnitId::Radian);
+        assert_eq!(out.unit, UnitId::Radian as u32);
     }
 
     #[test]
